+++ src/game/tests.rs
@@ -526,6 +526,7 @@ mod vintage_cube_six;
 mod vintage_cube_skyclave_apparition;
 mod vintage_cube_smugglers_copter;
 mod vintage_cube_spells;
+mod vintage_cube_springheart_nantuko;
 mod vintage_cube_staff_of_the_storyteller;
 mod vintage_cube_static_prison;
 mod vintage_cube_stoneforge;